pyproject.toml
[project]
name = "symbol-renamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyelftools",
    "lief",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello-world = "symbol_renamer:__main__"